workflows/requirements.txt
itis-dakota
# ipykernel
numpy
pandas 
# scipy
# matplotlib
# statsmodels